Trekke ut spesifikke hjemler til egne filer
