+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/brev/hjemler/OrdinæreHjemler.kt
@@ -1,0 +1,22 @@
+package no.nav.familie.ba.sak.kjerne.brev.hjemler
+
+import no.nav.familie.ba.sak.kjerne.vedtak.begrunnelser.hjemlerTilhørendeFritekst
+
+
+fun hentOrdinæreHjemler(
+    hjemler: MutableSet<String>,
+    opplysningspliktHjemlerSkalMedIBrev: Boolean,
+    finnesVedtaksperiodeMedFritekst: Boolean,
+): List<String> {
+    if (opplysningspliktHjemlerSkalMedIBrev) {
+        val hjemlerNårOpplysningspliktIkkeOppfylt = listOf("17", "18")
+        hjemler.addAll(hjemlerNårOpplysningspliktIkkeOppfylt)
+    }
+
+    if (finnesVedtaksperiodeMedFritekst) {
+        hjemler.addAll(hjemlerTilhørendeFritekst.map { it.toString() }.toSet())
+    }
+
+    val sorterteHjemler = hjemler.map { it.toInt() }.sorted().map { it.toString() }
+    return sorterteHjemler
+}
